+++ src/services/AssetTransforms.php
@@ -121,6 +121,18 @@ class AssetTransforms extends Component
     private $_activeTransformIndex;
 
     /**
+     * Serializer
+     *
+     * @since 3.5.14
+     */
+    public function __serialize()
+    {
+        $vars = get_object_vars($this);
+        unset($vars['_transforms']);
+        return $vars;
+    }
+
+    /**
      * @inheritdoc
      */
     public function init()