+++ src/services/Sections.php
@@ -133,6 +133,18 @@ class Sections extends Component
      */
     private $_entryTypesById;
 
+    /**
+     * Serializer
+     *
+     * @since 3.5.14
+     */
+    public function __serialize()
+    {
+        $vars = get_object_vars($this);
+        unset($vars['_sections']);
+        return $vars;
+    }
+
     // Sections
     // -------------------------------------------------------------------------